feat: pss-cln-def-fold-distribution-002 - Fold per-package asset/metadata modules with install materialize parity
Move goal, review, subagent, and tts asset/metadata implementation into distribution-owned private modules, leave transitional public shims, and reconcile ownership and coverage manifests while keeping goal decision-envelope interpretation in the public shim.
