+++ docs/internal/packaged-service-structure/package-target-manifest.json
@@ -112,11 +112,15 @@
     "pkg/services/factory_definitions/internal/services/compilation/wire",
     "pkg/services/factory_definitions/internal/services/distribution",
     "pkg/services/factory_definitions/internal/services/distribution/internal/service",
+    "pkg/services/factory_definitions/internal/services/distribution/goal",
     "pkg/services/factory_definitions/internal/services/distribution/packageassets",
     "pkg/services/factory_definitions/internal/services/distribution/packagedcatalog",
     "pkg/services/factory_definitions/internal/services/distribution/packagedinstallation",
     "pkg/services/factory_definitions/internal/services/distribution/promptassets",
+    "pkg/services/factory_definitions/internal/services/distribution/review",
     "pkg/services/factory_definitions/internal/services/distribution/scaffoldfacts",
+    "pkg/services/factory_definitions/internal/services/distribution/subagent",
+    "pkg/services/factory_definitions/internal/services/distribution/tts",
     "pkg/services/factory_definitions/internal/services/distribution/wire",
     "pkg/services/factory_definitions/internal/services/snapshots_portability",
     "pkg/services/factory_definitions/internal/services/snapshots_portability/capture",
@@ -918,6 +922,11 @@
       "destination": "factory_definitions/internal/services/distribution"
     },
     {
+      "packagePath": "pkg/services/factory_definitions/internal/services/distribution/goal",
+      "disposition": "retain",
+      "destination": "factory_definitions/internal/services/distribution"
+    },
+    {
       "packagePath": "pkg/services/factory_definitions/internal/services/distribution/packageassets",
       "disposition": "retain",
       "destination": "factory_definitions/internal/services/distribution"
@@ -938,7 +947,22 @@
       "destination": "factory_definitions/internal/services/distribution"
     },
     {
+      "packagePath": "pkg/services/factory_definitions/internal/services/distribution/review",
+      "disposition": "retain",
+      "destination": "factory_definitions/internal/services/distribution"
+    },
+    {
       "packagePath": "pkg/services/factory_definitions/internal/services/distribution/scaffoldfacts",
+      "disposition": "retain",
+      "destination": "factory_definitions/internal/services/distribution"
+    },
+    {
+      "packagePath": "pkg/services/factory_definitions/internal/services/distribution/subagent",
+      "disposition": "retain",
+      "destination": "factory_definitions/internal/services/distribution"
+    },
+    {
+      "packagePath": "pkg/services/factory_definitions/internal/services/distribution/tts",
       "disposition": "retain",
       "destination": "factory_definitions/internal/services/distribution"
     },
