+++ pkg/services/factory_definitions/internal/services/distribution/goal/prompt_paths.go
