+++ pkg/services/factory_definitions/internal/services/distribution/goal/summary.go
@@ -1,0 +1,8 @@
+package goal
+
+import "github.com/portpowered/infinite-you/pkg/services/factory_definitions/invocationoutput"
+
+var (
+	ShouldFormatInvocationSummary  = invocationoutput.ShouldFormatInvocationSummary
+	SummaryContentFromWorkerOutput = invocationoutput.SummaryContentFromWorkerOutput
+)
