+++ pkg/services/factory_definitions/internal/services/distribution/review/metadata.go
@@ -1,0 +1,10 @@
+package review
+
+const (
+	PackagedFactoryName                   = "@you/review"
+	PackagedFactoryProject                = "builtin-review"
+	PackagedWorkTypeName                  = "reviewable-work"
+	PackagedExecuteWorkstationName        = "execute-review-work"
+	PackagedReviewWorkstationName         = "review-review-work"
+	PackagedInvocationReturnTerminalState = "approved"
+)
