+++ pkg/services/factory_definitions/internal/services/distribution/subagent/response.go
@@ -1,0 +1,8 @@
+package subagent
+
+import "github.com/portpowered/infinite-you/pkg/services/factory_definitions/invocationoutput"
+
+var (
+	ShouldFormatInvocationResponse  = invocationoutput.ShouldFormatInvocationResponse
+	ResponseContentFromWorkerOutput = invocationoutput.ResponseContentFromWorkerOutput
+)
