+++ pkg/services/factory_definitions/packages/review/metadata_test.go
@@ -1,0 +1,6 @@
+package review_test
+
+import distributionreview "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/distribution/review"
+
+// Keep the public shim package exercised by delegating to the private owner tests.
+var _ = distributionreview.PackagedFactoryName
