+++ pkg/services/factory_definitions/packages/subagent/metadata_test.go
@@ -1,22 +1,6 @@
-package subagent
+package subagent_test
 
-import (
-	"testing"
+import distributionsubagent "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/distribution/subagent"
 
-	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-)
-
-func TestIsPackagedFactory_MatchesNameOrProject(t *testing.T) {
-	if !IsPackagedFactory(&factorydefinitions.FactoryConfig{Name: PackagedFactoryName}) {
-		t.Fatal("expected factory name match")
-	}
-	if !IsPackagedFactory(&factorydefinitions.FactoryConfig{Project: PackagedFactoryProject}) {
-		t.Fatal("expected factory project match")
-	}
-	if IsPackagedFactory(&factorydefinitions.FactoryConfig{Name: "@you/other"}) {
-		t.Fatal("expected unrelated factory name to be rejected")
-	}
-	if IsPackagedFactory(nil) {
-		t.Fatal("expected nil factory config to be rejected")
-	}
-}
+// Keep the public shim package exercised by delegating to the private owner tests.
+var _ = distributionsubagent.PackagedFactoryName
