+++ pkg/services/factory_definitions/packages/tts/metadata_test.go
@@ -1,58 +1,6 @@
-package tts
+package tts_test
 
-import (
-	"encoding/json"
-	"strings"
-	"testing"
+import distributiontts "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/distribution/tts"
 
-	"github.com/portpowered/infinite-you/pkg/services/work"
-	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-)
-
-func TestMetadataContentFromWorkerOutput_ReturnsTextMetadataWithoutRawAudio(t *testing.T) {
-	output := `[{"type":"AUDIO","file":"/tmp/speech.wav","contentType":"audio/wav","slot":"audio"}]`
-
-	got, err := MetadataContentFromWorkerOutput(output, "trace-tts", "~default", "")
-	if err != nil {
-		t.Fatalf("MetadataContentFromWorkerOutput: %v", err)
-	}
-	if len(got) != 1 || got[0].Type != work.WorkContentPartTypeText {
-		t.Fatalf("content = %#v, want one text metadata part", got)
-	}
-	if strings.Contains(got[0].Text, "AUDIO") {
-		t.Fatalf("metadata text = %q, want JSON metadata not raw audio content type marker", got[0].Text)
-	}
-
-	var metadata InvocationMetadata
-	if err := json.Unmarshal([]byte(got[0].Text), &metadata); err != nil {
-		t.Fatalf("metadata is not JSON: %v\n%s", err, got[0].Text)
-	}
-	if metadata.ArtifactPath != "/tmp/speech.wav" {
-		t.Fatalf("artifactPath = %q", metadata.ArtifactPath)
-	}
-	if metadata.MediaType != "audio/wav" {
-		t.Fatalf("mediaType = %q, want audio/wav", metadata.MediaType)
-	}
-	if metadata.Backend != "OMNIVOICE_Q4_K_M/LLAMACPP" {
-		t.Fatalf("backend = %q", metadata.Backend)
-	}
-	if metadata.TraceID != "trace-tts" || metadata.SessionID != "~default" {
-		t.Fatalf("metadata correlation = %#v", metadata)
-	}
-}
-
-func TestShouldFormatInvocationMetadata_MatchesPackagedInvokeWorkstation(t *testing.T) {
-	for _, workstationType := range []string{
-		factorydefinitions.WorkstationTypeInvoke,
-		factorydefinitions.WorkstationTypeInference,
-	} {
-		workstation := &factorydefinitions.FactoryWorkstationConfig{
-			Name:      PackagedInvokeWorkstationName,
-			Type:      workstationType,
-			Operation: "TTS",
-		}
-		if !ShouldFormatInvocationMetadata(workstation) {
-			t.Fatalf("expected packaged invoke workstation type %q to require metadata formatting", workstationType)
-		}
-	}
-}
+// Keep the public shim package exercised by delegating to the private owner tests.
+var _ = distributiontts.PackagedFactoryProject
